app/requirements.txt
aiobotocore==2.12.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
botocore==1.34.51
build==1.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
humanize==4.9.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.1.3
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pins==0.8.4
pip-tools==7.4.1
plotly==5.19.0
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsconnect_python==1.22.0
s3fs==2024.2.0
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==2.13.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tenacity==8.2.3
threadpoolctl==3.3.0
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.27.1
vetiver==0.2.4
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
